+++ packages/synapse-interface/components/Portfolio/SingleNetworkPortfolio.tsx
@@ -1,21 +1,26 @@
 import React, { useState } from 'react'
 import Image from 'next/image'
 import { CHAINS_BY_ID } from '@/constants/chains'
-import { TokenWithBalanceAndAllowance } from '@/utils/hooks/usePortfolioBalances'
+import {
+  ROUTER_ADDRESS,
+  TokenWithBalanceAndAllowance,
+  TokenWithBalanceAndAllowances,
+  separateTokensByAllowance,
+  sortTokensByBalanceDescending,
+} from '@/utils/actions/fetchPortfolioBalances'
 import { Chain } from '@/utils/types'
 import PortfolioAccordion from './PortfolioAccordion'
 import { PortfolioConnectButton } from './PortfolioConnectButton'
 import { EmptyPortfolioContent } from './PortfolioContent'
-import { FetchState } from '@/utils/hooks/usePortfolioBalances'
+import { FetchState } from '@/slices/portfolio/reducer'
 import { PortfolioTokenAsset } from './PortfolioTokenAsset'
 
 type SingleNetworkPortfolioProps = {
   portfolioChainId: number
   connectedChainId: number
   selectedFromChainId: number
-  portfolioTokens: TokenWithBalanceAndAllowance[]
+  portfolioTokens: TokenWithBalanceAndAllowances[]
   initializeExpanded: boolean
-  fetchPortfolioBalancesCallback: () => Promise<void>
   fetchState: FetchState
 }
 
@@ -25,7 +30,6 @@ export const SingleNetworkPortfolio = ({
   selectedFromChainId,
   portfolioTokens,
   initializeExpanded = false,
-  fetchPortfolioBalancesCallback,
   fetchState,
 }: SingleNetworkPortfolioProps) => {
   const currentChain: Chain = CHAINS_BY_ID[portfolioChainId]
@@ -33,12 +37,12 @@ export const SingleNetworkPortfolio = ({
   const [tokensWithAllowance, tokensWithoutAllowance] =
     separateTokensByAllowance(portfolioTokens)
 
-  const sortedTokensWithAllowance: TokenWithBalanceAndAllowance[] =
-    sortByBalanceDescending(tokensWithAllowance)
-  const sortedTokensWithoutAllowance: TokenWithBalanceAndAllowance[] =
-    sortByBalanceDescending(tokensWithoutAllowance)
-  const sortedTokensForVisualizer: TokenWithBalanceAndAllowance[] =
-    sortByBalanceDescending(portfolioTokens)
+  const sortedTokensWithAllowance: TokenWithBalanceAndAllowances[] =
+    sortTokensByBalanceDescending(tokensWithAllowance)
+  const sortedTokensWithoutAllowance: TokenWithBalanceAndAllowances[] =
+    sortTokensByBalanceDescending(tokensWithoutAllowance)
+  const sortedTokensForVisualizer: TokenWithBalanceAndAllowances[] =
+    sortTokensByBalanceDescending(portfolioTokens)
 
   const hasNoTokenBalance: boolean =
     !portfolioTokens || portfolioTokens.length === 0
@@ -75,15 +79,14 @@ export const SingleNetworkPortfolio = ({
         {sortedTokensWithAllowance &&
           sortedTokensWithAllowance.length > 0 &&
           sortedTokensWithAllowance.map(
-            ({ token, balance, allowance }: TokenWithBalanceAndAllowance) => (
+            ({ token, balance, allowances }: TokenWithBalanceAndAllowances) => (
               <PortfolioTokenAsset
                 key={token.symbol}
                 token={token}
                 balance={balance}
-                allowance={allowance}
+                allowances={allowances}
                 portfolioChainId={portfolioChainId}
                 connectedChainId={connectedChainId}
-                fetchPortfolioBalancesCallback={fetchPortfolioBalancesCallback}
                 isApproved={true}
               />
             )
@@ -91,14 +94,13 @@ export const SingleNetworkPortfolio = ({
         {sortedTokensWithoutAllowance &&
           sortedTokensWithoutAllowance.length > 0 &&
           sortedTokensWithoutAllowance.map(
-            ({ token, balance }: TokenWithBalanceAndAllowance) => (
+            ({ token, balance }: TokenWithBalanceAndAllowances) => (
               <PortfolioTokenAsset
                 key={token.symbol}
                 token={token}
                 balance={balance}
                 portfolioChainId={portfolioChainId}
                 connectedChainId={connectedChainId}
-                fetchPortfolioBalancesCallback={fetchPortfolioBalancesCallback}
                 isApproved={false}
               />
             )
@@ -137,7 +139,7 @@ const PortfolioNetwork = ({
 const PortfolioTokenVisualizer = ({
   portfolioTokens,
 }: {
-  portfolioTokens: TokenWithBalanceAndAllowance[]
+  portfolioTokens: TokenWithBalanceAndAllowances[]
 }) => {
   const [isHovered, setIsHovered] = useState(false)
   const hasOneToken = portfolioTokens && portfolioTokens.length > 0
@@ -181,7 +183,7 @@ const PortfolioTokenVisualizer = ({
             border border-solid border-[#252537]
             bg-[#101018] rounded-md`}
           >
-            {portfolioTokens.map((token: TokenWithBalanceAndAllowance) => {
+            {portfolioTokens.map((token: TokenWithBalanceAndAllowances) => {
               const tokenSymbol = token.token.symbol
               const balance = token.parsedBalance
               return (
@@ -209,38 +211,5 @@ export const PortfolioHeader = () => {
       </div>
       <div className="w-1/3 text-left" />
     </div>
-  )
-}
-
-function separateTokensByAllowance(
-  tokens: TokenWithBalanceAndAllowance[]
-): [TokenWithBalanceAndAllowance[], TokenWithBalanceAndAllowance[]] {
-  const tokensWithAllowance: TokenWithBalanceAndAllowance[] = []
-  const tokensWithoutAllowance: TokenWithBalanceAndAllowance[] = []
-
-  tokens &&
-    tokens.forEach((token) => {
-      // allowance is null for native gas tokens
-      if (token.allowance === null) {
-        tokensWithAllowance.push(token)
-      } else if (token.allowance > 0n) {
-        tokensWithAllowance.push(token)
-      } else {
-        tokensWithoutAllowance.push(token)
-      }
-    })
-
-  return [tokensWithAllowance, tokensWithoutAllowance]
-}
-
-function sortByBalanceDescending(
-  tokens: TokenWithBalanceAndAllowance[]
-): TokenWithBalanceAndAllowance[] {
-  return (
-    tokens &&
-    tokens.sort(
-      (a: TokenWithBalanceAndAllowance, b: TokenWithBalanceAndAllowance) =>
-        b.parsedBalance > a.parsedBalance ? 1 : -1
-    )
   )
 }