+++ packages/synapse-interface/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.69](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.68...@synapsecns/synapse-interface@0.1.69) (2023-07-24)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.68](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.67...@synapsecns/synapse-interface@0.1.68) (2023-07-21)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
 ## [0.1.67](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.66...@synapsecns/synapse-interface@0.1.67) (2023-07-20)
 
 **Note:** Version bump only for package @synapsecns/synapse-interface